Add example of a controller provider + class
1) controller.php: Mounted a prefix to the app with
a new instance of a controller provider class

2) ExampleControllerProvider: Implements Silex's
ControllerProviderInterface to provide routes for the controller
that will handle the requests coming in to the mounted prefix.
Instead of providing a closure for how the route should be handled,
we point at the class we want to use to handle the request.

3) ExampleController: Takes care of the rendering of the view